+++ src/pages/dataoffering/resources/resourceagreementsdialog/ResourceAgreementsDialog.vue
@@ -1,0 +1,3 @@
+<template src="./ResourceAgreementsDialog.html"></template>
+<script src="./ResourceAgreementsDialog.js"></script>
+<style src="./ResourceAgreementsDialog.css"></style>