+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project will be documented in this file.
 (Skipped major version 1, 2 and 3 to match versioning of IDS DataSpaceConnector)
 
+## [8.7.0] - 2021-12-17
+
+### Added
+- Filter offered resources by catalog
+- Set contract validity period for offered resources
+- Data Consumption: Show all resources instead of catalogs after request
+- Data Offering: Show agreements of resources
+
+### Changes
+- Use material icons in navigaiton menu
+
+### Fixes
+- Data Consumption: No validation on resource meta data dialog
+- Data Consumption: getIdOfAgreement() fixed
+
 ## [8.6.0] - 2021-10-08
 
 ### Added